+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 #### Added
 - Ignore and skip tables in automatic populator command
 - Add test combination foreign key and self foreign key to test class table analyzer
+- Support for PHP 8.0 and PHP 8.1
+
+#### Fixed
+- Progress bar
 
 ### [1.1.0] - 2021-05-26
 #### Added